requirements.txt
rpi_ws281x
adafruit-circuitpython-neopixel
adafruit-blinka
nicegui
webcolors
```